Stop token refresh causing empty bodies to be uploaded - Fixes #71
We try to use `Expect: 100-continue` if possible when doing http
transactions with a body.

If these fail with a 401 error the token can be refreshed and the
transaction retried.  If the transaction can't be retried we return a
swift.RetryNeeded error.

This also implements the DisableExpectContinue flag in
swift.Connection so using `Expect: 100-continue` can be turned off.
In this case any 401 errors returned while doing a transaction with a
body will return swift.RetryNeeded.